Merged PR 24656: SMS Sender cron script
## What's being changed

This PR adds a cron script to process the SMS order queue and send messages via the CPaas One API in batches.

## Why it's being changed

The `SmsSenderManager` is the core class - it handles the overall flow, delegating to specific classes to fetch items from the queue, batch images, and update rows after send.

When a batch is posted to the API, we receive a response with an array keyed according to the batch that was sent, and a message id for each message. We map this response back to the table, along with the phone number supplied for each message.

At the start of each run, we:

- expire any queue records if they are still pending after 24hrs
- check the status of each in-progress send (like the importer cron) and update the table accordingly
- process pending queue items

I've bundled a console command here. It's a sibling to the Email module's Sync command, and is extensible to trigger other non-sync processes which are also initiated via cron. This could be handy for the system alert Monitor cron.

## How to review / test this change

- Switch to branch `sms-develop` on the Email module
- Set up your local Magento to have at least two EC accounts connected to different websites
- Give one of your websites at least two store views
- Set different messages in your Transactional SMS message settings specific to store view e.g. 'SMS New order store 1', 'SMS New order store 2' etc.
- Seed your sms queue with some order sms rows corresponding to different stores and type ids
- Run the sender cron
- Test with phone numbers already in the table vs phone numbers pulled from the order shipping address
- Test queue items with corresponding orders that have no phone number, is the row marked as failed with a message?
- Check that the correct message text is sent for the row's type id
- Check that the correct message text is sent when processing the queue for two stores with different message content
- Check that processed items are marked as status 1 (in progress) with the correct message id. Look up the message id in Postman - does its phone number match the one in the table?

## Notes

- To run the new task runner CLI:
```
php bin/magento dotdigital:task SmsSenderManager
```

Related work items: #126407