Merged PR 28995: Use default error message if errorCode has no mapped message
## What's being changed

We're adding fallback error handling in case we don't have a mapped message for a particular validation error code.

## Why it's being changed

If the returned error code is -99 (and this happens in two potential cases), we don't have a mapped error text for this so we see a console error (in checkout) OR a message (in account) saying "Warning: No message defined for telephone".

## How to review / test this change

- Test in both account and checkout
- Start typing a number with 1-2 digits
- Click submit or click away (depending on context, you just want to trigger the validation)
- The error message should be 'Invalid telephone number'

Related work items: #142109