Merged PR 28992: Internationalize input value on blur
## What's being changed

We're adding an extra onBlur handler in each case that `intlTelInput` is initialised.

## Why it's being changed

Before this change it is possible that a user types a number in national mode (i.e. ignoring the placeholder to insert a country code, and not clicking the flag), and this passes validation, but is not valid for actually sending SMS.

By adding the onBlur we call `iti.getNumber()` when the customer has finished filling out their number. This will unobtrusively 'fix' the number and prepend a country code as per the flag.

## How to review / test this change

- Repeat these steps for checkout and account
- Land in the address edit screen
- Fill out a phone number without an international country code
- Watch the number fix itself when you click away or click submit
- Try to break it

Related work items: #141010