Merged PR 40429: Fall back to billing address where shipping address is not available
## What's being changed

We have updated the order SMS model to use the billing address phone number in cases where no shipping address phone number is provided (i.e. when purchasing downloadable and virtual products).

## Why it's being changed

Checking out with only virtual and/or downloadable products would result in a 500 error due to the lack of shipping information that was required to retrieve the customers contact number to send a sms order confirmation.

## How to review / test this change

1. Enable SMS and Order confirmation SMS
2. Add break point to `Dotdigitalgroup\Sms\Model\Queue\OrderItem\OrderItemNotificationEnqueuer@L71`
3. Checkout as a guest with only virtual and downloadable products in your cart
4. function call observe the $address object get assigned to billing address
5. you should be able to complete the order as expected.
6. Repeat the above using a shippable product and insure the shipping address is set not the billing address

## Notes

If you checkout as a customer with a previous order that contains shipping address the above mentioned is voided and would not count as a valid test case.

Related work items: #187933