Merged PR 40761: Check if customer is logged in before trigger validation for existing phone n...
## What's being changed

In this PR we check if a customer is logged in before we trigger an "existing mobile phone validation" in checkout page.

## Why it's being changed

Previously we added a feature that we provide an additional mobile input field in checkout step if the existing saved mobile phone number is in invalid format in order to allow logged in users to update their mobile phone number and then to receive sms as expected.

However it looks like we trigger this validation in guest checkout when guests filling their details via "autocomplete".
Especially in the case that a guest "autocompletes" his details with an invalid phone number (usually those details are saved in browser or google account fx) then we print out the additional mobile number form.
As a result we end up with two mobile number form fields.

## How to review / test this change

- Go to checkout page as a guest.
- Fill your details using autocomplete (and make sure you choose details with invalid phone number)
- Make sure that no additional mobile form is printed.

Related work items: #189019