Merged PR 44023: Add new columns to contact report
## What's being changed

The Contact Report has been modified to include new columns when the SMS module is enabled.

## Why it's being changed

We need to update the Contact Report so that merchants can see the status of contacts along with mobile numbers, subscriptions to the new SMS channel etc. We won't create a new report but instead use the same report and show all contacts, allowing merchants to use filters to create SMS-only views if they wish etc.

## How to review / test this change

1. Visit the Contact reports page Configuration > Dotdigital > Sync Settings >Sync > Customer Enabled > Click here for report
2. Insure you see the following columns
-- SMS Subscriber status
--  SMS Subscriber Imported
-- Mobile number
3. Modify the table `email_contact` columns `sms_subscriber_status`, `sms_subscriber_imported` and `mobile_number`
4. Observe the changes on the admin report page and insure the values reflect the story's acceptance criteria

Related work items: #202903, #204523, #204524