Merged PR 44693: Fix customer validation label positions
## What's being changed

We've fixed a validation snag on the customer register form - discovering in the process that the snag also affected customer account and customer address edit contexts.

We've also tightened up some of our JS templates, ensuring that they are only output in the DOM when they are actually needed.

## Why it's being changed

On submit, all validation messages relating to any field in the form were being added to our `mobile-number-error` div, thanks to our custom `errorLabelContainer` instruction. We introduced this to fix a positioning issue with the flag, but we've resolved that for Community with a CSS override, so we no longer need the `errorLabelContainer` instruction.

## How to review / test this change

- Retest validation in customer register and customer account, if SMS consent and phone number validation are switched on
- Retest the checkbox toggler UI, which should only be output if consent is enabled

Related work items: #209086