Merged PR 44501: Store SMS consent at registration, checkout, account area
## What's being changed

In this PR we added extra functionality in our consent plugins observers in order to create records in email_contact_consent table on registration, account area, checkout.

Also we added a new consent_text field in email_contact_consent table that captures the consent text by the time the contact subscribed to sms / email marketing.

## Why it's being changed

We need to capture any consent if the consent is enabled.

## How to review / test this change
- Enable Sms module and Enable sms consent in the Configuration > Dotdigital > Consent
- Make sure that a consent record is created on email_contact_consent table after you register an account (consent number provided), after updating the account area, on checkout (consent number needs to be provided too).
- Make sure that if no consent information provided in these steps then no record created,
- Make sure that when creating SMS consent records we capture the SMS consent text and not the Email consent.
- Disable Sms module or disable sms consent or disable both of them , repeat the above steps and make sure that no record created.

Related work items: #203938