Merged PR 44975: Align code to updated coding standards
## What's being changed

We have made several coding standards changes to the source, most of the changes are Dot block and accessibility declaration changes. We've also fixed a couple of unit tests for PHP 8.2.

## Why it's being changed

Get the module to pass our CI builds, in line with updates to coding standards and linting packages.

## How to review / test this change

1. Run PHPCS, PHPStan and unit tests
2. No errors should be present.

Related work items: #210416