Merged PR 45988: Add beforeSave hook to update sms_change_status_at
## What's being changed

An SmsContact model has been added, to extend from Dotdigitalgroup\Email\Model\Contact. In this class we have a `beforeSave` method that sets the new `sms_change_status_at` value whenever the `sms_subscriber_status` has changed.

## Why it's being changed

This is a convenient way to set the changed status at in one place. It is modelled on the core Newsletter Subscriber model.

## How to review / test this change

- Ensure that sms_change_status_at updates in all these cases:
  - Customer account subscribe
  - Customer account unsubscribe
  - Customer register
  - Customer order when consent is provided for SMS
  - Guest order when consent is provided for SMS
  - Admin customer subscribe
  - Admin customer unsubscribe

- Confirm that sms_change_status_at does NOT update when:
  - subscribing to the regular newsletter

## Notes
In line with our other models, we use Magento's internal DateTime class to set the 'now' time in the current timezone.

Related work items: #214637