Merged PR 46134: Validate phone number even if checkbox is ticked
## What's being changed

The changes in this PR mean we always trigger a validation for consent phone numbers, even if consent checkbox is NOT selected. The only time we do not validate is if a mobile phone number IS NOT present - meaning customers can still 'clear' their number.

## Why it's being changed

Because if checkbox is not ticked then we can save invalid phone numbers.

## How to review / test this change

- Go to my account area.
- Unselect sms consent checkbox.
- Try to save with an invalid phone number and make sure that validation error shows.
- Repeat the above steps by leaving the phone field empty.
- Confirm that you can save as expected.

Related work items: #215069