Merged PR 46148: Fix register page snag
## What's being changed

We've fixed a regression in our MarketingConsent view model.

## Why it's being changed

This view model is called from the marketing_consent.phtml template, which is used on the register page and the customer account page. It calls certain methods such as `getStoredMobileNumber` which is expected to return nothing when the customer is registering i.e. not logged in.

We modified this method recently and didn't test both bits of calling code.

## How to review / test this change

- Go to the register page - see no errors
- Sign in as a customer, go to your Subscriptions page - see your stored number loaded into the telephone input

Related work items: #216427