Merged PR 46315: Backup and restore SMS subscriber data during dotdigital:migrate
## What's being changed

This PR preserves SMS subscriber mobile numbers and subscriber status in the event that a merchant runs `dotdigital:migrate` after mobile numbers have already been captured.

At the start of the migration process - before the target tables are emptied - we store data for any contacts that have numbers, in a temporary table; then, after the other contact-type migrations have run, we update the new rows in `email_contact` with the stored data; and finally, after the whole process is complete, we drop the temporary table.

We only perform this backup-and-restore if the merchants intends to migrate ALL tables or ONLY the contact table.

## Why it's being changed

Post-install migration is occasionally used to repair data inconsistencies in our connector tables. Because `mobile_number` and `sms_subscriber_status` exist only in `email_contact`, running `dotdigital:migrate` would have meant losing stored numbers, and customers logging into their account and not seeing their previously stored number.

## How to review / test this change

- Set Customers > Customer configuration > Account sharing to 'per website'
- Start with an `email_contact` table that has many contacts with stored mobile numbers, ideally across multiple websites, some subscribed, some not subscribed
- Take an initial tally of the number of rows with numbers you expect to be backed up and restored
- Run `dotdigital:migrate` and observe the CLI output
- See the new steps added for backup and restore
- Do the rows affected match what you expected?
- Place multiple guest orders and set numbers at checkout - are these also restored?

_Global Account Sharing_
- Set Customers > Customer configuration > Account sharing to 'global'
- Register a customer on website 1 then log in to website 2, setting mobile numbers in both (giving you two rows in `email_contact`)
- Run `dotdigital:migrate`
- The row for website 2 will not be preserved
- Repeat the above steps, restoring a row for website 2
- Now place an order in website 2
- Again run `dotdigital:migrate`
- The row will be preserved

_Table restrictions_
- Run `dotdigital:migrate` with a `--table=email_contact` arg > see backup and restore steps
- Run `dotdigital:migrate` with a `--table=email_order` arg > do not see backup and restore steps

## Notes
If merchants specify that they want to migrate _only_ `email_contact`, we will back up and restore SMS subscribers, but this will not include subscriptions from guest sales. Those are migrated as part of the order migration. So, if those need to be preserved, the only option is to migrate everything.

Related work items: #211830