Merged PR 46881: Fix validation regression in customer account
## What's being changed

Our mobile number validation scripts.

## Why it's being changed

I spotted a regression in the telephone validation when adding a new customer address (the number only showed as invalid and the address could not be saved). However in addition I've removed some duplication around handing of the `countryCodeClass` (previously we would check if this was undefined twice - before the shared validation method was called, and inside it).

## How to review / test this change

- Test validation at registration
- Test validation in customer account (create / edit address)
- Test validation in customer account (SMS consent mobile number)
- Test validation at checkout (add / edit shipping phone number)
- Test validation at checkout (SMS consent mobile number)
- In all cases, test:
          - no number
          - too short number
          - invalid country code
          - missing flag (delete flag div from DOM to simulate)

## Notes

TelephoneValidatorUi has been updated, but the validate method isn't actually working in the admin customer edit view. This is a known limitation, we haven't been able to get validation working in this context.

Related work items: #220903