Merged PR 47404: Fix action group ref in TelephoneNumberPreValidationTest
## What's being changed

An action group reference in an MFTF test.

## Why it's being changed

This ref was incorrect, resulting in phone number validation not being switched off after the test was complete. When run in a bigger suite of Magento MFTF tests, this test breaks subsequent checkout tests because they are trying to validate when they don't need to.

## How to review / test this change

- Run `vendor/bin/mftf run:test TelephoneNumberPreValidationTest -r`
- Now run this core test: `vendor/bin/mftf run:test CheckCheckoutSuccessPageAsGuestTest -r`
- Both should succeed. You can also verify while the second test is running that the checkout phone number field is not rendered as an int tel input.

Related work items: #221647