Merged PR 49033: Use form.additional.info for SMS consent checkbox
## What's being changed

The layout for SMS blocks on the customer account register page.

## Why it's being changed

We had referenced a container `fieldset.create.info.additional` which only exists in the module Magento_LoginAsCustomerAssistance - so when that module is not active, or on Magento versions prior to 2.4.1, our checkbox does not display. Unfortunately it appears that this line in the register.phtml template:
```
<?= $block->getChildHtml('fieldset_create_info_additional') ?>
```
is a specific slot solely intended for the customer assistance module to use. `Magento_LoginAsCustomerAssistance` defines the `fieldset.create.info.additional` container, with an alias of `fieldset_create_info_additional`. So, when that module is active, no other module can reference that same name or fill that slot. It's the same with `customer.form.register.newsletter` (`$block->getChildHtml('customer.form.register.newsletter')` is an extension point but with only one extension in mind).

`form.additional.info` is different, because it is a container, and is declared in `customer_account_create.xml` - it's just not in the position we had it in originally.

## How to review / test this change

- Enable SMS consent
- Clear block_html and layout cache
- Go to storefront customer account create
- Confirm you see the SMS consent checkbox in the new location
- Run bin/magento module:disable Magento_LoginAsCustomerAssistance Magento_LoginAsCustomerGraphQl
- Run setup:upgrade
- Go to storefront customer account create
- Confirm you still see the SMS consent checkbox

Related work items: #231654