Merged PR 49654: Make sure that flag selector is loaded before checkout validation
## What's being changed

In this PR we are triggering the validation of the mobile form field in the checkout step once we've made sure that the required css classes are loaded in the dom.

## Why it's being changed

It looks like there is a case that when refreshing the checkout page that has already a number form filled that a validation is triggered but page breaks because at this point the elements we check js-side are not yet loaded.

## How to review / test this change

**Try to replicate**
- in develop branch go to checkout step and fill the form fields including our mobile form field.
- At that point **don't proceed to the next step** just refresh the page
- Confirm that blank page is loaded.

**Confirm that issue resolved**
- Switch to the branch
- Refresh the page again.
- Confirm that form displays correctly.
- try to replicate again from the beginning
- Please make sure that the validation of the phone field works as expected in all the states

Related work items: #234231