Merged PR 49505: Prevent missing order status from breaking checkout
## What's being changed

We are making some changes in light of reports from some merchants that elements of the Transactional SMS code is breaking checkout via Klarna Payments. See https://github.com/dotmailer/dotmailer-magento2-extension-sms/issues/7.

This PR:
- adds a failsafe check to all observers and plugins to exit first if Transactional SMS is not enabled
- relaxes the type requirements for setting additional data on our order item types, prior to queuing

## Why it's being changed

It appears that orders processed via Klarna Payments do not have an order status, and even if our Transactional SMS is not enabled, we are still setting data on an object that requires a status.

## How to review / test this change

- Test any selection of transactional SMS triggers
- Ensure that rows are still queued, even if additional data has blanks
- You can spoof the original reported problem by putting this in `Dotdigitalgroup\Sms\Model\Queue\OrderItem\NewOrder`:
```
$this->additionalData->setOrderStatus(null);
```

## Notes

Unfortunately I have not been able to reproduce this issue on staging with Klarna Payments configured (I have some token authorization issue I can't resolve). The main thing to check for end-to-end is, if data is null / not set / different from expected type, will our code throw an error.

Related work items: #233435