Merged PR 49353: Send transactional messages triggered by SMS opt-ins
## What's being changed

This PR triggers transactional SMS sends off the back of SMS opt-ins. We have one default message for this, and a separate one solely for new account opt-ins.

I've reused some bits of our existing transactional SMS queue system, but I've actually worked largely _alongside_ this because the existing system is very order-centric, and our new messages aren't tied solely to orders. We have two new queue item models and a different queuer class. I've renamed the table, and `order_id` is now nullable. On the messaging side, we now have two resolvers, one to deal with order-type messages, and another to handle customer-type messages. The connection between message types and their resolvers is set in di.xml.

Whenever mobile numbers are stored in either `email_contact` or `email_transactional_sms_queue`, we remove the leading '+', and restore it when retrieving the value. This is to ensure stored numbers match those returned by the Dotdigital API. This PR introduces a shared utility for this so we're consistent everywhere we store mobile numbers.

## Why it's being changed

This will allow welcome emails to be sent to newly opted in mobile numbers within a minute of opt-in.

## How to review / test this change

- Run setup:upgrade. Observe the changes to the database table.
- View the Transactional SMS admin, see two new templates with default message text and a list of available variables (note these are subject to change)
- Create a new customer, opting in to SMS consent > see a new row queued in `email_transactional_sms_queue` (type 7)
- In your account, go to the Subscriptions page
- Unsubscribe from SMS marketing, then re-subscribe > see a new row queued in `email_transactional_sms_queue` (type 6)
- Now place an order, opting in again to SMS marketing at checkout > see a new row queued in `email_transactional_sms_queue` (type 6)
- Run `bin/magento dotdigital:task SmsSenderManager` to test message sends and compilation with values as expected

## Notes

I made a change to the return value of SmsContact::setMobileNumber. Please include this in your testing. See commits on https://dev.azure.com/dotdigital/ec/_git/ec-core-magento2-extension-sms/pullrequest/45750?_a=commits for reference.

Related work items: #231941