Merged PR 49697: Retain reference to old email_sms_order_queue table
## What's being changed

db_schema.xml.

## Why it's being changed

Our `migrateDataFromAnotherTable` instruction, to change the table name from `email_sms_order_queue` to `email_transactional_sms_queue`, is not going to work for new installations of the module. The old table has to actually be present, and be active. So all we can do is mark the old one as deprecated, remove its indexes and foreign keys, and mark it as disabled at some point in the future.

## How to review / test this change

1.
- Observe the MFTF test build
- The install step must complete successfully

2.
- Delete your `email_transactional_sms_queue` table locally (take an export backup if you need)
- Delete your `email_sms_order_queue` table locally
- Run setup:upgrade
- Observe the table is restored, and that indexes and constraints are still in place with appropriate names

3.
- Check out to master branch
- Delete your `email_transactional_sms_queue` table
- Trigger an SMS item in the queue (putting an order on hold is quick)
- Switch back to this branch
- Run setup:upgrade
- Confirm that data is still migrated to the new table

4.
- Run
```
bin/magento setup:db-declaration:generate-whitelist --module-name=Dotdigitalgroup_Sms
```
- see no diff
- see no errors on CLI

Related work items: #234291