Merged PR 49939: Use context to get request instead of account_controller class
## What's being changed

Our Register observer for the `customer_register_success` event.

## Why it's being changed

This observer was relying on the `account_controller` arg being passed when the event is dispatched from `Magento\Customer\Controller\Account\CreatePost`. However it is possible for this event to be dispatched in other contexts without this argument. Safer to use `Magento\Framework\App\Action\Context`.

## How to review / test this change

- Turn on SMS consent
- Register a customer
- Confirm that an SMS subscription is stored in the `email_contact` table

Related work items: #235603