Merged PR 50570: Move cron job checker from cron to controller
## What's being changed

In line with similar work in the Email module, we've removed the job checker checks from the start of cron execution. This check now lives in the controller, so it checks for running jobs before starting a sync from the admin.

## Why it's being changed

This check was redundant (see explanation on Email PR).

## How to review / test this change

- Confirm cron jobs are scheduled as expected.
- Confirm SMS Subscriber sync can be run from the admin.

Related work items: #237941