Merged PR 53990: Block and layout updates to support Hyvä compatibility
## What's being changed

This PR adds some supporting CSS and templating code for our SMS Hyvä compatibility module. It also aligns the 'Subscriptions' menu amendment in the customer account with changes we've made to the Email module.

## Why it's being changed

As much as possible we should adapt code in the main module to be Hyvä-compatible, if it can support both, in order to minimise code in the compat module. The CSS addition is required because Hyvä has a slightly different folder structure when generating and outputting CSS into pub/static. We can add any Hyvä-specific style overrides here.

## How to review / test this change

- Confirm that the modified checkbox toggler still works in Luma
- Confirm that the 'Newsletter Subscriptions' menu item is re-labelled to 'Subscriptions' in the account (check Hyvä and Luma)

Related work items: #253324, #253624