Merged PR 55015: Consent prepopulation fix
## What's being changed

We have replaced the obscure `Object.entities` logic with the function `getInstance()`. We are now setting the value on the HTML element and the IntlInput wrapper.

## Why it's being changed

The phone validation for the consent at checkout was valid; however, the number was mutated by the intl input itself, causing the number to lose its dialing code "visually." This was causing unnecessary concern for the user of the module.

## How to review / test this change

1. Load up a new checkout form "Guest" with no pre-populated data.
2. Input a valid telephone number into the shipping phone field.
3. Check the consent box and observe an exact match of the shipping number.
4. Repeat this step with the exception of filling in the shipping number while logged in as a customer.

Related work items: #256006