Merged PR 55735: Consolidate SMS subscription state changes into a single queue and consumer
## What's being changed

The way we process SMS subscriptions via queues. We've consolidated subscription state changes so that they are handled sequentially in one consumer. Alongside this, there is now a single model for subscribe and unsubscribes.

## Why it's being changed

Using a single consumer, consuming subscription state changes on a single queue, feels like the best way to ensure that, for example, a subscribe is followed by an unsubscribe for a given contact. We have adopted a similar approach recently for newsletter subscriptions in the Email module.

## How to review / test this change

- Run setup:upgrade
- Test Checkout SMS subscription.
- Test Register SMS subscription.
- Test My Account area subscribe and unsubscribe.
- Test admin customer edit subscribe and unsubscribe.

Related work items: #260671