Merged PR 56572: Fix scope when queueing transactional SMS
## What's being changed

All points where we enqueue transactional SMS in the table. We're now obtaining the required, correct store id from the entity being observed, instead of using StoreManagerInterface.

## Why it's being changed

StoreManagerInterface does not give the correct store id in the Magento admin. We added these 'is enabled' checks in 2.2.0, but with this approach in a multi-store installation, creating a shipment from the order admin screen may find a different store id from the order being edited. This can result in messages not being queued as intended. The better approach is to take the store id in all cases from the order, or shipment, at hand.

## How to review / test this change

- Set up a multi-store Magento
- Configure Transactional SMS for website 2 and NOT at default or website 1 scopes
- Place an order on website 2 > see a message being queued
- In the Magento admin, change the status of the order > see another message queued
- Add a shipment > see another message queued
- Update this shipment with a new tracking number > see another message queued
- Put the order on hold > see another message queued
- Complete the order and issue a credit memo > see another message queued

Related work items: #264371