Merged PR 58235: Update ConsentGuestCheckoutValidationTest MFTF
## What's being changed

We're updating an MFTF test that had started failing following [this commit](47a6a4cbe8c7b650bd353852d5e64ed79868940f).

## Why it's being changed

We had previously altered the expected phone number value in the consent input, because of the way it used to work (i.e. different from all the other inputs). Now that we fixed the underlying reason for why it behaved this way, the test can be changed to what it should have been in the first place.

## How to review / test this change

- See a PASS for this test

Related work items: #271479