Merged PR 58300: Add try catch when publishing to queues
## What's being changed

We've wrapped any calls to PublisherInterface::publish in try/catch statements. In addition, all this module's observers are now wrapped in try/catch.

One other small thing - the observers now return `$this` as opposed to void. This doesn't change anything, because Magento doesn't do anything with the return value of an observer - but it's a convention from core that we stick to everywhere else.

## Why it's being changed

It's important that observers and plugins in modules must never derail active processes in Magento. The `publish` call does throw a LocalizedException if the name of the topic is not recognised.

## How to review / test this change

- Register as a customer, setting SMS consent with a phone number
- Confirm the number is saved
- In the admin, go the customer record and subscribe / unsubscribe the customer's mobile for SMS marketing
- Confirm both actions complete as expected
- Change the topic names for each to a bad topic name and confirm the thrown exception is caught, logged, and does not surface an error to the storefront

Related work items: #265574