Merged PR 59357: Remove overloaded migration method
## What's being changed

The `bulkUpdate` method has been removed in the `RestoreEmailContactTableSmsSubscribers` data migration.

## Why it's being changed

This method was only present here in order to pass an array to `getUpdateBindings()` instead of a key string, which is what the parent did. I've now changed the parent to take an array, so this overloading is no longer required.

AbstractBulkUpdater was initially created to update a specific column in the target table with a named key from the select. Then `RestoreEmailContactTableSmsSubscribers` came along and needed to update 3 columns. Hence the need for an array.

## How to review / test this change

- Take a note of the count for this migration type before switching branches
- Switch to this branch, re-run migrate, and check the count matches

Related work items: #240661