Merged PR 59532: Remove keywords as originators in transactional SMS
## What's being changed

The list of possible originators returned in the source model for 'Select default SMS from name'.

I also updated the `getFromNumber` method to make it more agnostic (it can return a name or a number, and it was confusing to have a `getFromNumber` method that returned a variable `$fromName`).

## Why it's being changed

Keywords are unlikely to be used for transactional SMS sends, so we're removing them from the source model to reduce noise.

## How to review / test this change

- Switch to AM DD account
- See dedicated number and shortcode in list of options for 'Select default SMS from name'
- Set the long number as your originator
- Enable transactional SMS for e.g. new order
- Place an order (use a UK number)
- See SMS delivered successfully

Related work items: #276468