Merged PR 60336: Add backend models to prevent consent being enabled without consent text
## What's being changed

Enabling consent for SMS marketing in the connector admin. Merchants must now set consent text when enabling consent. In addition, merchants cannot clear consent text if consent is enabled in any context.

## Why it's being changed

Consent text is a requirement for a valid consent record. Our SDK requires this and throws an error, which can disrupt both consent sync and fetching data from the API, as well as saves in connector/customer/index. So the way to prevent all of these is to prevent the text being empty in the first place. Resolves https://github.com/dotmailer/dotmailer-magento2-extension/issues/629 .

## How to review / test this change

- Go to Dotdigital > Marketing Consent
- Test enabling SMS marketing consent with and without text in either textarea
- Confirm this works at website and store scope and with inherited values
- Set everything up so it's correct - then try deleting consent text while consent is Yes in any context or at any level (this should also not be possible)

Related work items: #216620