Merged PR 58505: Fix MFTF test annotation errors
## What's being changed

Fixes some annotation errors flagged in the BAP report. Also I've added a line to `ConsentGuestCheckoutValidationDisabledTest` which we have on `ConsentGuestCheckoutValidationTest`, for consistency.

## Why it's being changed

The line addition relating to Flat Rate doesn't appear to fix the occasional errors we get in `ConsentGuestCheckoutValidationTest` and `ConsentGuestCheckoutValidationDisabledTest` on the step "selectFlatRateShippingMethod", it's mainly for consistency.

The Adobe report said:
> The test failed on step "selectFlatRateShippingMethod"; it looks like the step "waitForConsentMobileNumber" needs more time to load.
> To fix this, add <wait> time or <see userInput="$5.00 Fixed Flat Rate" selector=../> before  <click selector="{{CheckoutShippingMethodsSection.checkShippingMethodByName('Flat Rate')}}" stepKey="selectFlatRateShippingMethod"/>

I added this, and it still tripped up at the same point.

## How to review / test this change

- See a green build in MFTF

Related work items: #275301