Merged PR 67273: Fix console errors for SMS consent
## What's being changed

We have added a conditional check to ensure the element exists before attempting to retrieve it from the Intl input globals object.

## Why it's being changed

The console was throwing an error during runtime if the element had not loaded by the time the intlInput wanted to attached itself to the element in question causing the fetch to fail.

## How to review / test this change

- Enable SMS consent
- Open console on frontend.
- Log in as a customer
- Proceed to shipping with a saved address
- Observe no console errors.

Related work items: #306842