Merged PR 67587: Stop loading the telephone input at checkout when validation is disabled
## What's being changed

This PR prevents the IntlTelInput loading at checkout if telephone validation is disabled.

## Why it's being changed

There was a regression some time back connected to a refactor for Hyvä themes. The impact is that telephone input validation would block submission / checkout even if the merchant had chosen not to use our validator.

## How to review / test this change

- Enable SMS consent to be captured at checkout
- Test the consent input with validation enabled
- Test the consent input with validation disabled
- See a green build for all checkout-related MFTF tests

Related work items: #310380