Merged PR 69563: Modify ConsentRegistrationValidationTest to test validation
## What's being changed

This PR modifies an MFTF test to pass reliably. Instead of looking for the error message inside the `.mage-error` div, we now just confirm that the input is marked with the class `.mage-error`.

## Why it's being changed

At some point since we initially wrote this test it started to fail regularly. Seemingly when run in a headless test, the validation error messages don't appear. I've reproduced this when running the test locally. It only seems to affect this particular form (we have similar tests in the customer account and the checkout). I've used interactive pause, and run the test to the point when it loads the form. I've tested it during the pause, and it works - just not when the test is left to run by itself.

## How to review / test this change

- See a green build

Related work items: #310588