Merged PR 71010: Remove dotdigital.com email check for PAYG notice
## What's being changed

This removes a check we had in place to allow local enablement of Transactional SMS even if our target accounts did not officially have a paid PAYG flag for sending SMS.

## Why it's being changed

Initially when we set this up, we didn't allow merchants to turn on the feature if they weren't set up for Pay-As-You-Go SMS sending. Later we relaxed this restriction, but we kept the notice in place. However we continued to hide it for internal @dotdigital.com email accounts, meaning that we never saw the notice. I've removed the check now because I think it's more helpful to be reminded that the notice exists and will display.

## How to review / test this change

- Go to the Transactional SMS section
- See the PAYG notice when switching on Transactional SMS

Related work items: #325781